Create Login.py
This UDF, or rather, collection of UDFs, will be used to create users and save progress, as well as access traineR itself, and will be used as the UDF for option [2] of introChoice in morseL.

I will admit that this code is fraught with errors (or at least, according to VS Code; I see no problems at all apart from `cred()` ) , and weeding them out will probably take some time and effort from the final editor's end, but I've spent a lot of time developing the code so far, and we can fix it up later.

cred() does not exist yet due to SQL refusing to cooperate with me, so I'll have to depend on you for that. Again, as I have mentioned before, if you need any help with bugs, I'll do my best to be of help. As I explained, cred() should open the SQL database and check the login credentials input by the user and return whether they are correct or not in the form of 
`return True` or `return False`